trustee-cli: add subject alternative name
Add a Subject Alternative Name to the self-signed certificate,
the kbs-client failed without it.

Always overwrite the self-signed certificate. Needed to avoid the
complexity of refreshing the certificate if the listening IP addresses
change or if the certificate expires. Add "cache" to the certificate
file name to convey that the file has no lifetime guarantees.

Signed-off-by: Camilla Conte <cconte@redhat.com>